API: add dtype= option to python parser
