+++ pandas/io/tests/parser/test_parsers.py
@@ -22,6 +22,7 @@ from .parse_dates import ParseDatesTests
 from .compression import CompressionTests
 from .multithread import MultithreadTests
 from .python_parser_only import PythonParserTests
+from .dtypes import DtypeTests
 
 
 class BaseParser(CommentTests, CompressionTests,
@@ -29,7 +30,8 @@ class BaseParser(CommentTests, CompressionTests,
                  IndexColTests, MultithreadTests,
                  NAvaluesTests, ParseDatesTests,
                  ParserTests, SkipRowsTests,
-                 UsecolsTests, QuotingTests):
+                 UsecolsTests, QuotingTests,
+                 DtypeTests):
     def read_csv(self, *args, **kwargs):
         raise NotImplementedError
